backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-modeltranslation==0.18.11
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
Pillow==10.1.0
python-decouple==3.8
django-filter==23.5
drf-spectacular==0.26.5
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0